refactor(memory-cli): split IDE platform code into memory/ide submodule
memory.rs had grown to 1774 lines mixing memory CRUD with IDE detection,
MCP config installs, filesystem/env abstraction, and ~700 lines of
detection tests. Split the IDE concern out into a sibling module and
collapse the per-IDE duplication along the way.

- Move AiTool, DetectionEnv/RealEnv, detection probes, install_*,
  LoadedConfig, InstallOutcome, run_mcp_install, and all detection tests
  into fluree-db-cli/src/commands/memory/ide.rs. memory.rs now declares
  `mod ide;`, delegates Phase 2 of init via `ide::run_mcp_phase(yes)`,
  and routes McpInstall to `ide::run_mcp_install`.
- Make IDE detection data-driven via `IdeDetectionSpec` (binary, macOS
  app, home markers, config markers). One `is_ide_present(env, spec)`
  replaces five near-identical `is_*_present` functions; `path_exists`
  matches both files and dirs, collapsing Claude's `.claude` (dir) vs
  `.claude.json` (file) into one marker list. Adding another IDE is now
  a table edit. Already-configured probes are similarly table-driven via
  `already_configured_probes(tool, env)`.
- Extract `install_json_mcp_config(JsonMcpInstall { config_path, top_key,
  default, entry })` — the load/merge/refuse-on-unsafe/write shape used
  by Cursor / VS Code / Windsurf / Zed. Per-IDE installers shrink to a
  few lines each.
- Extract `install_rules_file(dir)` for the bundled fluree_rules.md.
  Best-effort: prints a warning on failure instead of propagating, so a
  rules-file IO error no longer masks a successful MCP config install in
  the per-tool summary.
- Add `build_synced_store(dirs)` to memory.rs, replacing 7 copy-pasted
  `build_store + ensure_synced` pairs in the CRUD subcommands. run_init
  keeps a plain build_store since it creates an empty store before sync.
- Rename `InstallOutcome::Skipped` → `ManualRequired` and
  `LoadedConfig::Default` → `MissingOrEmpty` for more precise intent.
  The init summary now reports "N tools need manual setup" instead of
  "Skipped N tools".
- Tighten load_config: a parsed JSON value that isn't an object now
  returns Unsafe, since merge_mcp_entry can't merge into arrays /
  scalars / null. Previously a top-level `[]` would have been written
  back as-is with InstallOutcome::Installed reported to the user.

Tests: 28 unit tests in commands::memory::ide::tests (up from 21),
covering project-root walk-up, per-IDE presence via each signal class,
Claude already-configured under nested cwd, the load_config matrix
(missing, empty, JSONC, invalid UTF-8, non-object root, valid),
install_json_mcp_config (happy + refuse + non-object), and the
best-effort install_rules_file failure path. CLI test suite: 73/73.

memory.rs is now 498 lines focused on dispatch + memory CRUD; ide.rs is
1305 lines and self-contained.
